Move quotas in another package
Move quotas in selvpcclient/quotamanager.
Add mutex in identity manager.
Separate quotas client and requests.
