+++ selvpcclient/quotamanager/client.go
@@ -1,4 +1,4 @@
-package quotas
+package quotamanager
 
 import (
 	"context"
