+++ selvpcclient/quotamanager/doc.go
@@ -1,0 +1,4 @@
+/*
+Package quotamanager provides methods and structures to work with the Quotas Manager API.
+*/
+package quotamanager
